+++ .envrc
@@ -1,0 +1,2 @@
+dotenv
+env_vars_required APP_HOST
