+++ .gitignore
@@ -1,3 +1,4 @@
 dist/
 target/
 test_keys/
+.env
