+++ src/components/byte_input.rs
@@ -1,8 +1,8 @@
 use web_sys::HtmlInputElement;
 use yew::{classes, function_component, html, use_effect_with_deps, use_state, Callback, Html, Properties, TargetCast};
 
-use super::BytesFormat;
-use crate::common::{encode_bytes, get_format_button_class, get_set_format_callback, parse_bytes, BYTES_FORMATS};
+use crate::common::ui::{get_format_button_class, get_set_format_callback};
+use crate::common::{encode_bytes, parse_bytes, BytesFormat, BYTES_FORMATS};
 
 #[derive(PartialEq, Properties, Clone)]
 pub struct ByteInputProps {
