+++ src/components/info/algo_search.rs
