+++ src/components/input/bcrypt.rs
@@ -2,7 +2,9 @@ use web_sys::HtmlInputElement;
 use yew::{classes, function_component, html, Callback, Html, Properties, TargetCast};
 use yew_notifications::{use_notification, Notification, NotificationType};
 
-use crate::common::{build_byte_input, BytesFormat, Switch};
+use crate::common::BytesFormat;
+use crate::components::byte_input::build_byte_input;
+use crate::components::switch::Switch;
 use crate::crypto_helper::algorithm::{BcryptAction, BcryptHashAction, BcryptInput as BcryptInputData};
 
 #[derive(PartialEq, Properties, Clone)]
