+++ src/jwt/mod.rs
@@ -1,0 +1,7 @@
+#[allow(clippy::module_inception)]
+pub mod jwt;
+pub mod jwt_utils;
+pub mod jwte;
+pub mod signature;
+#[macro_use]
+mod macros;
