+++ src/routes/mod.rs
@@ -1,0 +1,9 @@
+mod about;
+mod crypto_helper;
+mod jwt;
+mod not_found;
+
+pub use about::About;
+pub use crypto_helper::CryptoHelper;
+pub use jwt::Jwt;
+pub use not_found::not_found;
