+++ src/settings.rs
@@ -1,0 +1,1 @@
+pub const APP_HOST: &str = std::env!("APP_HOST");
